refactor(team): split launcher_lifecycle.go into three phase files (C20)
launcher_lifecycle.go (386 lines) bundled three semantically distinct
phases of the Launcher's runtime. Splitting into one file per phase
makes each file's responsibility legible at a glance and lets the
reader find a method by what it does, not by the catch-all name
"lifecycle":

* launcher_boot.go (171 lines, new) — Launch(): the tmux-mode boot
  path. Composes the broker, spawns the channel TUI in tmux, kicks
  off the headless context + notification goroutines, and returns
  once the session is steady-state. Web-mode boot lives in
  launcher_web.go (LaunchWeb); headless-only boot lives in
  headless_codex.go (launchHeadlessCodex). All three share the
  broker.Start + writeOfficePIDFile + goroutine fan-out shape but
  differ in whether they spawn tmux or print a web URL.

* launcher_session.go (97 lines, new) — user-facing session
  state-change methods: Attach (re-attach the user's terminal to the
  running tmux session), Kill (graceful drain + tear-down),
  ResetSession (clear broker state + manifest + per-agent temp files
  for a fresh-start). Each is small (15-30 lines) but together they
  form the "user can drive the running team" surface.

* launcher_reconfigure.go (157 lines, new) — the runtime-
  reconfiguration flow. ReconfigureSession + reconfigureVisibleAgents
  + respawnPanesAfterReseed handle the case where the user adds/
  removes agents mid-session: re-resolve LLM provider, decide pane
  vs headless, tear down + respawn against new roster while
  preserving the channel pane and pane sizes when possible.

launcher_lifecycle.go deleted; the three phase files take its place.

Cumulative since main: launcher.go 4998 -> 332 (-93.4%).

Stacked on refactor/launcher-c19-misc-cleanup (C19).
